A priori feature selection, transformation with transcan
Abortive attempt to use subject matter to pre-determine predictors (as the number of data points is too few for data driven feature selection). This was successful.
However, also attempted to use transcan to transform data to optimise distribution of predictors and attempted to transform the dentists and dent_personnel predictors to minimise collinearity, but both transformations led to increased collinearity and poorer model performance